+++ cmd/octopool/string_rewrites_attachments_open_other.go
@@ -1,0 +1,9 @@
+//go:build !aix && !android && !darwin && !dragonfly && !freebsd && !illumos && !ios && !linux && !netbsd && !openbsd && !solaris && !windows
+
+package main
+
+import "os"
+
+func openRewriteAttachment(path string) (*os.File, error) {
+	return os.Open(path)
+}