+++ cmd/octopool/string_rewrites_attachments_open_unix.go
@@ -1,0 +1,12 @@
+//go:build aix || android || darwin || dragonfly || freebsd || illumos || ios || linux || netbsd || openbsd || solaris
+
+package main
+
+import (
+	"os"
+	"syscall"
+)
+
+func openRewriteAttachment(path string) (*os.File, error) {
+	return os.OpenFile(path, os.O_RDONLY|syscall.O_NONBLOCK, 0)
+}